+++ packages/@webex/plugin-meetings/src/meeting/index.ts
@@ -121,17 +121,17 @@ import ControlsOptionsManager from '../controls-options-manager';
 
 const {isBrowser} = BrowserDetection();
 
-const logRequest = (request: any, {header = '', success = '', failure = ''}) => {
-  LoggerProxy.logger.info(header);
+const logRequest = (request: any, {logText = ''}) => {
+  LoggerProxy.logger.info(`${logText} - sending request`);
 
   return request
     .then((arg) => {
-      LoggerProxy.logger.info(success);
+      LoggerProxy.logger.info(`${logText} - has been successfully sent`);
 
       return arg;
     })
     .catch((error) => {
-      LoggerProxy.logger.error(failure, error);
+      LoggerProxy.logger.error(`${logText} - has failed: `, error);
       throw error;
     });
 };
@@ -523,6 +523,7 @@ export default class Meeting extends StatelessWebexPlugin {
   state: any;
   localAudioTrackMuteStateHandler: (event: TrackMuteEvent) => void;
   localVideoTrackMuteStateHandler: (event: TrackMuteEvent) => void;
+  webexMeetingId: string;
 
   namespace = MEETINGS;
 
@@ -559,7 +560,7 @@ export default class Meeting extends StatelessWebexPlugin {
      */
     this.id = uuid.v4();
     /**
-     * Correlation ID used for network tracking of meeting join
+     * Correlation ID used for network tracking of meeting
      * @instance
      * @type {String}
      * @readonly
@@ -616,7 +617,7 @@ export default class Meeting extends StatelessWebexPlugin {
      * @memberof Meeting
      */
     // @ts-ignore
-    this.breakouts = new Breakouts({}, {parent: this.webex});
+    this.breakouts = new Breakouts({meetingId: this.id}, {parent: this.webex});
     /**
      * helper class for managing receive slots (for multistream media connections)
      */
@@ -2321,6 +2322,9 @@ export default class Meeting extends StatelessWebexPlugin {
           canStartTranscribing: MeetingUtil.canStartTranscribing(payload.info.userDisplayHints),
           canStopTranscribing: MeetingUtil.canStopTranscribing(payload.info.userDisplayHints),
           isClosedCaptionActive: MeetingUtil.isClosedCaptionActive(payload.info.userDisplayHints),
+          isSaveTranscriptsEnabled: MeetingUtil.isSaveTranscriptsEnabled(
+            payload.info.userDisplayHints
+          ),
           isWebexAssistantActive: MeetingUtil.isWebexAssistantActive(payload.info.userDisplayHints),
           canViewCaptionPanel: MeetingUtil.canViewCaptionPanel(payload.info.userDisplayHints),
           isRealTimeTranslationEnabled: MeetingUtil.isRealTimeTranslationEnabled(
@@ -2589,6 +2593,12 @@ export default class Meeting extends StatelessWebexPlugin {
         },
         EVENT_TRIGGERS.MEETING_BREAKOUTS_UPDATE
       );
+    });
+
+    // We need to reinitialize  when user upgrades to host or cohost
+    this.locusInfo.on(LOCUSINFO.EVENTS.SELF_MODERATOR_OR_COHOST_UPGRADE, (payload) => {
+      this.breakouts.queryPreAssignments(payload);
+      // ...
     });
 
     this.locusInfo.on(LOCUSINFO.EVENTS.SELF_IS_SHARING_BLOCKED_CHANGE, (payload) => {
@@ -3512,8 +3522,6 @@ export default class Meeting extends StatelessWebexPlugin {
       return Promise.reject(new ParameterError('no audio control associated to the meeting'));
     }
 
-    const LOG_HEADER = 'Meeting:index#muteAudio -->';
-
     // First, stop sending the local audio media
     return logRequest(
       this.audio
@@ -3537,9 +3545,7 @@ export default class Meeting extends StatelessWebexPlugin {
           throw error;
         }),
       {
-        header: `${LOG_HEADER} muting audio`,
-        success: `${LOG_HEADER} muted audio successfully`,
-        failure: `${LOG_HEADER} muting audio failed, `,
+        logText: `Meeting:index#muteAudio --> correlationId=${this.correlationId} muting audio`,
       }
     );
   }
@@ -3565,8 +3571,6 @@ export default class Meeting extends StatelessWebexPlugin {
       return Promise.reject(new ParameterError('no audio control associated to the meeting'));
     }
 
-    const LOG_HEADER = 'Meeting:index#unmuteAudio -->';
-
     // First, send the control to unmute the participant on the server
     return logRequest(
       this.audio
@@ -3590,9 +3594,7 @@ export default class Meeting extends StatelessWebexPlugin {
           throw error;
         }),
       {
-        header: `${LOG_HEADER} unmuting audio`,
-        success: `${LOG_HEADER} unmuted audio successfully`,
-        failure: `${LOG_HEADER} unmuting audio failed, `,
+        logText: `Meeting:index#unmuteAudio --> correlationId=${this.correlationId} unmuting audio`,
       }
     );
   }
@@ -3618,8 +3620,6 @@ export default class Meeting extends StatelessWebexPlugin {
       return Promise.reject(new ParameterError('no video control associated to the meeting'));
     }
 
-    const LOG_HEADER = 'Meeting:index#muteVideo -->';
-
     return logRequest(
       this.video
         .handleClientRequest(this, true)
@@ -3642,9 +3642,7 @@ export default class Meeting extends StatelessWebexPlugin {
           throw error;
         }),
       {
-        header: `${LOG_HEADER} muting video`,
-        success: `${LOG_HEADER} muted video successfully`,
-        failure: `${LOG_HEADER} muting video failed, `,
+        logText: `Meeting:index#muteVideo --> correlationId=${this.correlationId} muting video`,
       }
     );
   }
@@ -3670,8 +3668,6 @@ export default class Meeting extends StatelessWebexPlugin {
       return Promise.reject(new ParameterError('no audio control associated to the meeting'));
     }
 
-    const LOG_HEADER = 'Meeting:index#unmuteVideo -->';
-
     return logRequest(
       this.video
         .handleClientRequest(this, false)
@@ -3694,9 +3690,7 @@ export default class Meeting extends StatelessWebexPlugin {
           throw error;
         }),
       {
-        header: `${LOG_HEADER} unmuting video`,
-        success: `${LOG_HEADER} unmuted video successfully`,
-        failure: `${LOG_HEADER} unmuting video failed, `,
+        logText: `Meeting:index#unmuteVideo --> correlationId=${this.correlationId} unmuting video`,
       }
     );
   }
@@ -4821,7 +4815,7 @@ export default class Meeting extends StatelessWebexPlugin {
     this.mediaProperties.webrtcMediaConnection.on(Event.ROAP_FAILURE, this.handleRoapFailure);
 
     this.mediaProperties.webrtcMediaConnection.on(Event.ROAP_MESSAGE_TO_SEND, (event) => {
-      const LOG_HEADER = 'Meeting:index#setupMediaConnectionListeners.ROAP_MESSAGE_TO_SEND -->';
+      const LOG_HEADER = `Meeting:index#setupMediaConnectionListeners.ROAP_MESSAGE_TO_SEND --> correlationId=${this.correlationId}`;
 
       switch (event.roapMessage.messageType) {
         case 'OK':
@@ -4837,9 +4831,7 @@ export default class Meeting extends StatelessWebexPlugin {
               correlationId: this.correlationId,
             }),
             {
-              header: `${LOG_HEADER} Send Roap OK`,
-              success: `${LOG_HEADER} Successfully send roap OK`,
-              failure: `${LOG_HEADER} Error joining the call on send roap OK, `,
+              logText: `${LOG_HEADER} Roap OK`,
             }
           );
           break;
@@ -4859,9 +4851,7 @@ export default class Meeting extends StatelessWebexPlugin {
               reconnect: this.reconnectionManager.isReconnectInProgress(),
             }),
             {
-              header: `${LOG_HEADER} Send Roap Offer`,
-              success: `${LOG_HEADER} Successfully send roap offer`,
-              failure: `${LOG_HEADER} Error joining the call on send roap offer, `,
+              logText: `${LOG_HEADER} Roap Offer`,
             }
           );
           break;
@@ -4880,9 +4870,7 @@ export default class Meeting extends StatelessWebexPlugin {
               correlationId: this.correlationId,
             }),
             {
-              header: `${LOG_HEADER} Send Roap Answer.`,
-              success: `${LOG_HEADER} Successfully send roap answer`,
-              failure: `${LOG_HEADER} Error joining the call on send roap answer, `,
+              logText: `${LOG_HEADER} Roap Answer`,
             }
           ).catch((error) => {
             const metricName = BEHAVIORAL_METRICS.ROAP_ANSWER_FAILURE;
@@ -4919,9 +4907,7 @@ export default class Meeting extends StatelessWebexPlugin {
               correlationId: this.correlationId,
             }),
             {
-              header: `${LOG_HEADER} Send Roap Error.`,
-              success: `${LOG_HEADER} Successfully send roap error`,
-              failure: `${LOG_HEADER} Failed to send roap error, `,
+              logText: `${LOG_HEADER} Roap Error (${event.roapMessage.errorType})`,
             }
           );
           break;
@@ -5015,7 +5001,7 @@ export default class Meeting extends StatelessWebexPlugin {
       };
 
       LoggerProxy.logger.info(
-        `Meeting:index#setupMediaConnectionListeners --> connection state changed to ${event.state}`
+        `Meeting:index#setupMediaConnectionListeners --> correlationId=${this.correlationId} connection state changed to ${event.state}`
       );
       switch (event.state) {
         case ConnectionState.Connecting: